home_automation/requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
cryptography==2.3.1
emoji==0.5.1
future==0.16.0
idna==2.7
pycparser==2.19
PySocks==1.6.8
python-telegram-bot==11.1.0
RPi.GPIO==0.6.4
six==1.11.0
